Refactor and use original colors
