+++ tailwind.config.js
@@ -5,7 +5,14 @@ module.exports = {
   content: [...themeConfig.content, './src/**/*.{js,jsx,ts,tsx}'],
   theme: {
     ...themeConfig.theme,
+    extend: {
+      ...themeConfig.theme.extend,
+      colors: {
+        gray: { hover: '#40364d' }
+      }
+    },
     fontFamily: {
+      ...themeConfig.theme.fontFamily,
       sans: ['BrandonGrotesque', 'Tahoma', 'Arial', 'sans-serif'],
       mono: ['Consolas', '"Liberation Mono"', 'Menlo', 'Courier', 'monospace']
     }
